Moved mutate to utils
